% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{add_one}
\alias{add_one}
\title{Add one}
\usage{
add_one(x)
}
\arguments{
\item{x}{A vector}
}
\value{

}
\description{
Add one to some number or numbers
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
