% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{\%c\%}
\alias{\%c\%}
\title{Compute regression coefficients}
\usage{
y \%c\% x
}
\arguments{
\item{y}{\code{vector} of responses}

\item{x}{design \code{matrix}}
}
\description{
Compute regression coefficients
}
\details{
Compute the regression coefficients using \link[stats]{lm}.
}
\examples{
## linear regression
cars$speed \%c\% cars$dist
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{coef}}
}
