% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{\%r\%}
\alias{\%r\%}
\title{Function to calculate the regression coefficients}
\usage{
y \%r\% x
}
\arguments{
\item{y}{The dependent variable}

\item{x}{The independent variable}
}
\value{
The regression coefficients
}
\description{
This function calculates the regression coefficients of a linear model
}
\examples{
## linear regression
cars$speed \%r\% cars$dist
}
