% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello world function}
\usage{
hello()
}
\value{
print a message
}
\description{
hello world function
}
\details{
A super fancy function to print `Hello World!`
}
\examples{
\dontrun{hello()}
}
\author{
John Doe
}
