% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInstaller.R
\name{ipak}
\alias{ipak}
\title{ipak Function}
\usage{
ipak(pkg)
}
\arguments{
\item{pkg}{user-created list of packages, concatenated and individually quoted}
}
\description{
This function hecks to see if packages are installed, installs them if they are not, then loads them into the R session..
}
\examples{
ipak()
}
\keyword{packages}
