% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstfunction.R
\name{mean_by}
\alias{mean_by}
\title{to generate the summary dataframe}
\usage{
mean_by(data, by, var)
}
\arguments{
\item{data}{input the dataframe}

\item{by}{input the by group or groups}

\item{var}{input the variable}
}
\value{
a dataframe
}
\description{
to generate the summary dataframe
}
\examples{
mean_by(mtcars, rlang::exprs(cyl, vs, am), mpg)

}
