% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgTest-package.R
\docType{package}
\name{pkgTest-package}
\alias{pkgTest-package}
\alias{pkgTest}
\title{The 'pkgTest' package.}
\description{
A minital package using STAN code
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
