% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{regression_coefficient}
\alias{regression_coefficient}
\title{Compute regression coefficients}
\usage{
regression_coefficient(x, y)
}
\arguments{
\item{x}{design \code{matrix}}

\item{y}{\code{vector} of responses}
}
\description{
Compute regression coefficients
}
\details{
Compute the regression coefficients using \link[stats]{lm}.
}
\examples{
## linear regression
regression_coefficient(x = cars$speed, y = cars$dist)
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{coef}}
}
