% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIW.R
\name{rIW}
\alias{rIW}
\title{Random generation from IW distribution}
\usage{
rIW(n, d, R)
}
\arguments{
\item{n}{number of observations}

\item{d}{degrees of freedom IW parameter}

\item{R}{matrix IW parameter}
}
\description{
Random generation from IW distribution
}
\examples{
rIW(n=4, d = 3, R = diag(2) )

}
