% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{\%r\%}
\alias{\%r\%}
\title{Compute regression coefficients}
\usage{
y \%r\% x
}
\arguments{
\item{y}{\code{vector} of responses}

\item{x}{design \code{matrix}}
}
\description{
Compute regression coefficients
}
\details{
Compute the regression coefficients using \link[stats]{lm}.
}
\examples{
## linear regression
cars$speed \%r\% cars$dist
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{coef}}
}
