% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello world function}
\usage{
hello()
}
\value{
Print the "Hello, world!" message
}
\description{
This function print a simple message.
The goal is to test R package development.
}
\author{
John Doe
}
